requirements_copter.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fqdn==1.5.1
fsspec==2024.9.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
MAVProxy==1.8.71
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.2
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.2
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pymavlink==2.4.41
pynmeagps==1.0.43
pyrtcm==1.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyubx2==1.2.47
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scipy==1.14.1
screeninfo==0.8.1
Send2Trash==1.8.3
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
uri-template==1.3.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
